% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{blaze_version}
\alias{blaze_version}
\title{Get blaze version}
\usage{
blaze_version()
}
\value{
A string with the blaze version name and number
}
\description{
Provides the blaze C++ library version number included in the package.
}
\examples{
blaze_version()
}
