% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11blaze-package.R
\docType{package}
\name{cpp11blaze-package}
\alias{cpp11blaze}
\alias{cpp11blaze-package}
\title{cpp11blaze: A 'Blaze' Interface}
\description{
Provides function declarations and inline function definitions that facilitate communication between R and the 'Blaze' 'C++' library for linear algebra and scientific computing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11blaze/}
  \item \url{https://github.com/pachadotdev/cpp11blaze}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11blaze/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item  BLAZE AUTHORS (Blaze library (C++))
}

}
\keyword{internal}
